Migrate ArgoCD operator profile integration test to Java
